Two Aliens Abby & Gabby Adventure on Earth Episode
Vasitum is an AI-based HRtech startup, focusing on solving the challenges of recruiters and job seekers alike. The platform removes all the heavy-lifting of the placement process. AI-recommendations based on 90% match rates cut out the difficulty of search. Paired with automated real-time communication, screening  and scheduling, Vasitum is a modern-day HR-powerhouse to shorten the employment process and fosters everyone’s ability to reach their full potential. 

For more information visit - https://vasitum.com/